feat: add dev enviroment
